# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Fleet Intelligence Copilot",
    "description": "Synthetic fleet operations with cited policies, bounded tools, human approvals, and auditable actions. No paid services required.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/copilot/ask": {
      "post": {
        "tags": [
          "Copilot"
        ],
        "summary": "Ask",
        "operationId": "ask_v1_copilot_ask_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AskRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AskResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/v1/actions/propose": {
      "post": {
        "tags": [
          "Approval workflow"
        ],
        "summary": "Propose",
        "operationId": "propose_v1_actions_propose_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProposalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/v1/approvals/{approval_id}": {
      "get": {
        "tags": [
          "Approval workflow"
        ],
        "summary": "Approval",
        "operationId": "approval_v1_approvals__approval_id__get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "approval_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Approval Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/approvals/{approval_id}/decision": {
      "post": {
        "tags": [
          "Approval workflow"
        ],
        "summary": "Decide",
        "operationId": "decide_v1_approvals__approval_id__decision_post",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "approval_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Approval Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DecisionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/approvals/{approval_id}/execute": {
      "post": {
        "tags": [
          "Approval workflow"
        ],
        "summary": "Execute",
        "operationId": "execute_v1_approvals__approval_id__execute_post",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "approval_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Approval Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AskRequest": {
        "properties": {
          "question": {
            "type": "string",
            "maxLength": 1000,
            "minLength": 3,
            "title": "Question"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "question"
        ],
        "title": "AskRequest"
      },
      "AskResponse": {
        "properties": {
          "trace_id": {
            "type": "string",
            "title": "Trace Id"
          },
          "intent": {
            "type": "string",
            "title": "Intent"
          },
          "risk": {
            "type": "string",
            "title": "Risk"
          },
          "status": {
            "type": "string",
            "enum": [
              "answered",
              "needs_clarification",
              "blocked",
              "degraded"
            ],
            "title": "Status"
          },
          "answer": {
            "type": "string",
            "title": "Answer"
          },
          "data": {
            "additionalProperties": true,
            "type": "object",
            "title": "Data"
          },
          "citations": {
            "items": {
              "$ref": "#/components/schemas/Citation"
            },
            "type": "array",
            "title": "Citations"
          },
          "suggested_action": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Suggested Action"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "trace_id",
          "intent",
          "risk",
          "status",
          "answer"
        ],
        "title": "AskResponse"
      },
      "Citation": {
        "properties": {
          "policy_id": {
            "type": "string",
            "title": "Policy Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "section": {
            "type": "string",
            "title": "Section"
          },
          "version": {
            "type": "string",
            "title": "Version"
          },
          "excerpt": {
            "type": "string",
            "title": "Excerpt"
          },
          "source": {
            "type": "string",
            "title": "Source"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "policy_id",
          "title",
          "section",
          "version",
          "excerpt",
          "source"
        ],
        "title": "Citation"
      },
      "DecisionRequest": {
        "properties": {
          "decision": {
            "type": "string",
            "enum": [
              "approve",
              "reject"
            ],
            "title": "Decision"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "decision"
        ],
        "title": "DecisionRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ProposalRequest": {
        "properties": {
          "action": {
            "type": "string",
            "enum": [
              "support_case",
              "escalation",
              "replacement_recommendation"
            ],
            "title": "Action"
          },
          "vehicle_id": {
            "type": "string",
            "pattern": "^V-\\d{3}$",
            "title": "Vehicle Id"
          },
          "idempotency_key": {
            "type": "string",
            "maxLength": 80,
            "minLength": 8,
            "pattern": "^[\\w-]+$",
            "title": "Idempotency Key"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "action",
          "vehicle_id",
          "idempotency_key"
        ],
        "title": "ProposalRequest"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  }
}